Added alt-text to SVG logo links
From Trello task "Add alt-text to the Facebook and Twitter website links on the website". Added aria-labelledby attribute and <title> tags to <svg> in header and footer for Facebook and Twitter logo links. Removed aria-hidden attribute (false is the default) to allow content to be presented to screen reader users.